requirements.txt
algoliasearch==2.6.3
algoliasearch-django==2.0.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
certifi==2022.12.7
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.6
Django==4.0.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.9.0
fonttools==4.49.0
idna==3.6
inflection==0.5.1
install==1.3.5
ipykernel==6.29.3
ipython==8.22.1
jedi==0.19.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mlxtend==0.23.1
nest-asyncio==1.6.0
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
openpyxl==3.1.1
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
pipenv==2023.2.18
pippip==0.0.10
platformdirs==3.0.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
pygame==2.5.2
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2022.7.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.4
stack-data==0.6.3
threadpoolctl==3.3.0
tornado==6.4
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.19.0
virtualenv-clone==0.5.7
wcwidth==0.2.13
xlrd==2.0.1
